Modularizacion de algunos componentes y mejora en la logica de algunas funciones
