--- src/config/index.js
@@ -1,9 +1,0 @@
-export const SCHEDULE = "schedules";
-export const USER = "users";
-export const SERVICE = "products";
-export const QUOTE = "quotes";
-export const AUTH  = {
-    AUTH: "auth",
-    LOGOUT: "logout",
-    CHECKSESSION: "check-session"
-};
