+++ src/config/schedule.js
@@ -1,105 +1,37 @@
-import {SCHEDULE} from "@/config/index";
-import toast from "react-hot-toast";
+import {SCHEDULE, fetchWithErrorHandling} from "@/config/constants";
 
-export const getSchedules = async() => {
-    try {
-        const response = await fetch(`${process.env.NEXT_PUBLIC_SERVER_HOST}/${SCHEDULE}`, {
-            method: 'GET',
-            headers: {
-                'Content-Type': 'application/json',
-            },
 
-            credentials: "include"
-        });
-
-        const result = await response.json();
-        if(result?.error) {
-            toast.error(result?.error);
-            return null;
-        };
-
-        return result;
-    } catch (error) {
-        console.log("getSchedules: ", error);
-        return false;
-    };
-};
-
-export const putSchedule = async(id, data) => {
-    try {
-        
-        const response = await fetch(`${process.env.NEXT_PUBLIC_SERVER_HOST}/${SCHEDULE}`, {
-            method: 'PUT',
-            headers: {
-                'Content-Type': 'application/json',
-            },
-            body: JSON.stringify({ id, ...data }),
-            credentials: "include"
-        });
-        
-        const result = await response.json();
-        
-        if(result?.error) {
-            toast.error(result.error);
-            return null;
-        };
-
-        return result;
-    } catch (error) {
-        console.log("postSchedule: ", error);
-        return false;
-    };
-};
-
-export const deleteShedule = async(id) => {
-    try {
-        
-        const response = await fetch(`${process.env.NEXT_PUBLIC_SERVER_HOST}/${SCHEDULE}`, {
-            method: 'DELETE',
-            headers: {
-                'Content-Type': 'application/json',
-            },
-            body: JSON.stringify({ id }),
-            credentials: "include"
-        });
-        
-        const result = await response.json();
-        
-        if(result?.error) {
-            toast.error(result.error);
-            return null;
-        };
-
-        return result;
-    } catch (error) {
-        console.log("postSchedule: ", error);
-        return false;
-    };
-};
-
-export const postSchedule = async(startTime, endTime, status, day, month, year) => {
-    try {
-        
-        const response = await fetch(`${process.env.NEXT_PUBLIC_SERVER_HOST}/${SCHEDULE}`, {
-            method: 'POST',
-            headers: {
-                'Content-Type': 'application/json',
-            },
-            body: JSON.stringify({ startTime, endTime, status, day, month, year }),
-            credentials: "include"
-        });
-        
-        const result = await response.json();
-    
-        if(result?.error) {
-            toast.error(result.error);
-            return null;
-        };
-
-        return result;
-    } catch (error) {
-        console.log("postSchedule: ", error);
-        return false;
-    };
-};
-
+// Obtener Horarios
+export const getSchedules = async () => {
+    return await fetchWithErrorHandling(`${process.env.NEXT_PUBLIC_SERVER_HOST}/${SCHEDULE}`, {
+      method: 'GET',
+      headers: { 'Content-Type': 'application/json' },
+    });
+  };
+  
+  // Editar horario
+  export const putSchedule = async (id, data) => {
+    return await fetchWithErrorHandling(`${process.env.NEXT_PUBLIC_SERVER_HOST}/${SCHEDULE}`, {
+      method: 'PUT',
+      headers: { 'Content-Type': 'application/json' },
+      body: JSON.stringify({ id, ...data }),
+    });
+  };
+  
+  // Eliminar horario
+  export const deleteSchedule = async (id) => {
+    return await fetchWithErrorHandling(`${process.env.NEXT_PUBLIC_SERVER_HOST}/${SCHEDULE}`, {
+      method: 'DELETE',
+      headers: { 'Content-Type': 'application/json' },
+      body: JSON.stringify({ id }),
+    });
+  };
+  
+  // Crear horario
+  export const postSchedule = async (data) => {
+    return await fetchWithErrorHandling(`${process.env.NEXT_PUBLIC_SERVER_HOST}/${SCHEDULE}`, {
+      method: 'POST',
+      headers: { 'Content-Type': 'application/json' },
+      body: JSON.stringify(data),
+    });
+  };
