+++ src/hooks/useSchedules.js
@@ -1,39 +1,52 @@
 import { useUser } from "@/config/UserContext";
-import { useState, useEffect } from "react";
-import { getSchedules, putSchedule, deleteShedule } from"@/config/schedule";
+import { useState, useEffect, useCallback } from "react";
+import { getSchedules, putSchedule, deleteSchedule, postSchedule } from "@/config/schedule";
 import { useSocket } from "@/config/socketContext";
 
-export const useSchedules = () => {
+export const useSchedules = (initialSchedules) => {
     const { user } = useUser();
     const socketClient = useSocket();
-    const [schedules, setSchedules] = useState([]);
-    
-    const requestShedules = async () => { 
-        await getSchedules();  
-    };
+    const [schedules, setSchedules] = useState(initialSchedules || []);
 
-    const editSchedule = async(idSchedule, data) => {
+    // Función para manejar la obtención de horarios
+    const requestSchedules = useCallback(async () => {
+        await getSchedules();
+    }, []);
+
+    // Manejo de edición de horarios
+    const editSchedule = useCallback(async (idSchedule, data) => {
         const result = await putSchedule(idSchedule, data);
         return result;
-    }
+    }, []);
 
-    const dltSchedule = async(idSchedule) => {
-        const result = await deleteShedule(idSchedule);
+    // Manejo de eliminación de horarios
+    const dltSchedule = useCallback(async (idSchedule) => {
+        const result = await deleteSchedule(idSchedule);
         return result;
-    };
+    }, []);
+
+    // Manejo de creación de horarios
+    const addSchedule = useCallback(async (data) => {
+        const result = await postSchedule(data);
+        return result;
+    }, []);
 
     useEffect(() => {
-        if(user) {
-            socketClient.on("getScheduleList", (schedules)=> {
+        if (user) {
+            socketClient.on("getScheduleList", (schedules) => {
                 setSchedules(schedules || []);
             });
 
-            if(schedules.length === 0) {
-                requestShedules();
+            if (schedules.length === 0) {
+                requestSchedules();
+            };
+
+            return () => {
+                socketClient.off("getScheduleList");
             };
         };
-    
+
     }, [user, socketClient]); // El arreglo vacío [] asegura que esto solo se ejecute al montar/desmontar el componente
 
-    return { schedules, editSchedule, dltSchedule };
+    return { schedules, editSchedule, dltSchedule, addSchedule, addSchedule };
 };
